requirements.txt
asgiref==3.8.1
certifi==2024.6.2
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
defusedxml==0.7.1
Django
django-cors-headers==4.4.0
django-environ==0.11.2
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
gunicorn==22.0.0
httpie==3.2.2
idna==3.7
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
numpy
razorpay
oauthlib==3.2.2
opencv-python==4.10.0.84
packaging==24.1
pillow==10.3.0
psycopg2-binary==2.9.9
pycparser==2.22
Pygments==2.18.0
PyJWT==2.8.0
PySocks==1.7.1
python-decouple==3.8
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
setuptools==70.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.0
urllib3==2.2.1
whitenoise==6.7.0
accelerate==0.25.0
torchmetrics==1.2.1
tqdm==4.66.1
transformers==4.36.2
diffusers==0.25.0
einops==0.7.0
bitsandbytes==0.39.0
scipy
opencv-python
gradio==4.24.0
fvcore
cloudpickle
omegaconf
pycocotools
basicsr
av
onnxruntime==1.16.2

